+++ swing/common-model/src/test/java/is/codion/swing/common/model/component/table/DefaultTableSelectionTest.java
@@ -99,7 +99,7 @@ public class DefaultTableSelectionTest {
 		assertEquals(1, emptyCounter.get());
 		testModel.addSelectionInterval(0, 1);
 		assertEquals(1, emptyCounter.get());
-		testModel.indexes().moveDown();
+		testModel.indexes().increment();
 		assertEquals(1, emptyCounter.get());
 		testModel.clearSelection();
 		assertEquals(2, emptyCounter.get());